Fixed issues with Space Engineers Script Studio.
