+++ SpaceEngineersScriptStudio/SpaceEngineersScriptStudio.sln
@@ -1,22 +1,22 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
-VisualStudioVersion = 16.0.32407.337
+VisualStudioVersion = 16.0.32901.82
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpaceEngineersScripting", "SpaceEngineersScripting\SpaceEngineersScripting.csproj", "{0461BBD4-0AF9-4971-B1D8-8680EC5337BD}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpaceEngineersScriptStudio", "SpaceEngineersScriptStudio\SpaceEngineersScriptStudio.csproj", "{1221D588-2B38-4F15-9DBC-1FEC2CF25912}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0461BBD4-0AF9-4971-B1D8-8680EC5337BD}.Debug|x64.ActiveCfg = Debug|x64
-		{0461BBD4-0AF9-4971-B1D8-8680EC5337BD}.Debug|x64.Build.0 = Debug|x64
+		{1221D588-2B38-4F15-9DBC-1FEC2CF25912}.Debug|x64.ActiveCfg = Debug|x64
+		{1221D588-2B38-4F15-9DBC-1FEC2CF25912}.Debug|x64.Build.0 = Debug|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8E741216-1693-4D2E-B9D6-37992E4F5684}
+		SolutionGuid = {0D289974-F596-4073-8213-5FB778B45070}
 	EndGlobalSection
 EndGlobal
