+++ SpaceEngineersScriptStudio/SpaceEngineersScriptStudio/Instructions.readme
