+++ SpaceEngineersScriptStudio/SpaceEngineersScriptStudio/SpaceEngineersScriptStudio.csproj
@@ -4,10 +4,10 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{0461BBD4-0AF9-4971-B1D8-8680EC5337BD}</ProjectGuid>
+    <ProjectGuid>{1221D588-2B38-4F15-9DBC-1FEC2CF25912}</ProjectGuid>
     <OutputType>Exe</OutputType>
     <RootNamespace>IngameScript</RootNamespace>
-    <AssemblyName>SpaceEngineersScripting</AssemblyName>
+    <AssemblyName>SpaceEngineersScriptStudio</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
