+++ SpaceEngineersScriptStudio/SpaceEngineersScriptStudio/thumb.png
